Exposed capability and person hierarchy rebuilds via jmx
 #108
