+++ waltz-jobs/src/main/java/com/khartec/waltz/jobs/CapabilityHarness.java
@@ -17,10 +17,6 @@
 
 package com.khartec.waltz.jobs;
 
-import com.khartec.waltz.common.hierarchy.Forest;
-import com.khartec.waltz.common.hierarchy.Node;
-import com.khartec.waltz.model.capability.Capability;
-import com.khartec.waltz.model.capability.ImmutableCapability;
 import com.khartec.waltz.service.DIConfiguration;
 import com.khartec.waltz.service.capability.CapabilityService;
 import com.khartec.waltz.service.capability_usage.CapabilityUsageService;
@@ -33,47 +29,10 @@ public class CapabilityHarness {
 
         AnnotationConfigApplicationContext ctx = new AnnotationConfigApplicationContext(DIConfiguration.class);
 
-        CapabilityUsageService capabilityUsageService = ctx.getBean(CapabilityUsageService.class);
         CapabilityService capabilityService = ctx.getBean(CapabilityService.class);
 
-        capabilityService.assignLevels();
-
-        Forest<Capability, Long> forest = capabilityService.buildHierarchy();
-
-        for (Node<Capability, Long> node : forest.getAllNodes().values()) {
-            Capability capability = node.getData();
-            int currLevel = capability.level();
-
-            Node ptr = node;
-            while (ptr != null) {
-                capability = setLevel(capability, currLevel, ptr);
-                ptr = ptr.getParent();
-                currLevel--;
-            }
-
-            capabilityService.update(capability);
-        }
+        capabilityService.rebuildHierarchy();
 
     }
-
-    private static Capability setLevel(Capability capability, int level, Node<Capability, Long> node) {
-        if (level == 1) {
-            return ImmutableCapability.copyOf(capability).withLevel1(node.getId());
-        }
-        if (level == 2) {
-            return ImmutableCapability.copyOf(capability).withLevel2(node.getId());
-        }
-        if (level == 3) {
-            return ImmutableCapability.copyOf(capability).withLevel3(node.getId());
-        }
-        if (level == 4) {
-            return ImmutableCapability.copyOf(capability).withLevel4(node.getId());
-        }
-        if (level == 5) {
-            return ImmutableCapability.copyOf(capability).withLevel5(node.getId());
-        }
-        return capability;
-    }
-
 
 }
